moved associations related logic in to adapter
